+++ tests/test_config/test_views.py
@@ -4175,7 +4175,9 @@ class DummyRegistry:
     utility = None
 
     def __init__(self):
-        self.settings = {}
+        self.settings = {
+            'pyramid.reload_assets': False,
+        }
 
     def queryUtility(self, type_or_iface, name=None, default=None):
         return self.utility or default